requirements.txt
streamlit
numpy
torch
opencv-contrib-python-headless
timm
Pillow==8.0.0
torchvision
pytorch_lightning
matplotlib